Extract nested property methods into dedicated files
Move nested-related code out of shared files into focused companions:
- inverted/objects_nested.go: index flag helpers (isNested*, HasNested*,
  HasAnyNestedInvertedIndex), nestedIndexConfig, collectNestedIndexConfig,
  analyzeNestedProp, analyzeNestedValue
- inverted/objects_nested_test.go: all TestHasNested*, TestCollectNestedIndexConfig,
  TestAnalyzeNestedProp tests
- shard_init_properties_nested.go: createNestedPropertyBuckets
- shard_write_inverted_lsm_nested.go: extendNestedInvertedIndicesLSM,
  extendNestedFilterableIndex, extendNestedMetaIndex

Also reorder objects_nested.go so related methods are adjacent:
predicates (isNested*) → Has* helpers → nestedIndexConfig + collect → analysis.
