+++ adapters/repos/db/inverted/analyzer.go
@@ -58,13 +58,13 @@ type NestedValue struct {
 // NestedProperty holds all analyzed values and metadata from position
 // assignment of a single nested property.
 type NestedProperty struct {
-	Name               string         // top-level property name (for bucket naming)
-	Values             []NestedValue  // analyzed values for the value bucket
-	Idx                []NestedMeta   // _idx metadata entries
-	Exists             []NestedMeta   // _exists metadata entries
-	HasFilterableIndex bool           // any value needs the filterable bucket
-	HasSearchableIndex bool           // any value needs the searchable bucket
-	HasRangeableIndex  bool           // any value needs the rangeable bucket
+	Name               string        // top-level property name (for bucket naming)
+	Values             []NestedValue // analyzed values for the value bucket
+	Idx                []NestedMeta  // _idx metadata entries
+	Exists             []NestedMeta  // _exists metadata entries
+	HasFilterableIndex bool          // any value needs the filterable bucket
+	HasSearchableIndex bool          // any value needs the searchable bucket
+	HasRangeableIndex  bool          // any value needs the rangeable bucket
 }
 
 // NestedMeta is an _idx or _exists metadata entry from position assignment.
